Modules: geometry, my_DSP_modules, my_operations, antenna_array
geometry.py:

Defines classes and methods for geometry related usage. It consists
of class "Point" which defines a point in 3D space and includes
operator overloading.

my_operations.py:

Defines operations not available under usual python libraries.
It consists of reverse() which reverses a given axis in a
multi-dimensional data set (up to 8 dimensions) and for the given
subset of indices along this axis.

my_DSP_modules.py:

Defines operations usually associated with FFT. The exact FFT
procedure is chosen based on whether the data is real or complex.
It performs a 1D FFT on a multi-dimensional data set (up to 8
dimensions) and computes the frequencies in the computed spectrum
besides some dimensional juggling acts.

antenna_array.py:

Defines "Antenna" and "Antenna_Pair" as classes. "Antenna" stores
electric field data arriving at that antenna besides the antenna
label. It also has a Fourier Transform built in as one of its methods.
"Antenna_Pair" defnes the class for a pair of antennas. Its attributes
includes the time series of electric fields, the cross-correlation
function of the electric fields between the two antennas, the Fourier
transform of the cross-correlation which is computed using FX() and
XF() approaches. These are included as the class methods.